Move sync watchdog snapshot and tests
