+++ sync/src/lib.rs
@@ -6,6 +6,7 @@ pub mod announcement;
 pub mod block_connector;
 pub mod store;
 pub mod sync;
+pub mod sync_watchdog;
 pub mod sync_metrics;
 pub mod tasks;
 pub mod txn_sync;
