+++ sync/src/sync_watchdog.rs
@@ -1,0 +1,63 @@
+// Copyright (c) The Starcoin Core Contributors
+// SPDX-License-Identifier: Apache-2.0
+
+use std::time::{Duration, Instant};
+
+pub const SYNC_WATCHDOG_INTERVAL: Duration = Duration::from_secs(30);
+pub const SYNC_WATCHDOG_STALL_SECS: u64 = 120;
+
+#[derive(Clone, Debug)]
+pub struct SyncWatchdogSnapshot {
+    pub task_name: String,
+    pub processed: u64,
+    pub ok: u64,
+    pub last_change: Instant,
+}
+
+pub fn update_watchdog_state(
+    state: &mut Option<SyncWatchdogSnapshot>,
+    task_name: String,
+    processed: u64,
+    ok: u64,
+    now: Instant,
+) -> bool {
+    match state.as_mut() {
+        Some(snapshot)
+            if snapshot.task_name != task_name
+                || processed < snapshot.processed
+                || ok < snapshot.ok =>
+        {
+            *snapshot = SyncWatchdogSnapshot {
+                task_name,
+                processed,
+                ok,
+                last_change: now,
+            };
+            false
+        }
+        Some(snapshot) => {
+            if processed > snapshot.processed || ok > snapshot.ok {
+                snapshot.processed = processed;
+                snapshot.ok = ok;
+                snapshot.last_change = now;
+                return false;
+            }
+            if now.duration_since(snapshot.last_change).as_secs() >= SYNC_WATCHDOG_STALL_SECS {
+                snapshot.last_change = now;
+                snapshot.processed = processed;
+                snapshot.ok = ok;
+                return true;
+            }
+            false
+        }
+        None => {
+            *state = Some(SyncWatchdogSnapshot {
+                task_name,
+                processed,
+                ok,
+                last_change: now,
+            });
+            false
+        }
+    }
+}
